Add files via upload
A C7 cut contains light hydrocarbons dissolved in non-negligible quantities: they are mainly C2 with a little methane and some C3. It is envisaged to eliminate these light constituents by distillation (flash or stripping). 

A preliminary study of the operating conditions of the process can be done in pseudo-binary: we assimilate the C7 cut to n-heptane and the light ones to ethane. We wish to construct the diagrams [T-x-y] and [x-y], [h-x-y] of the ethane-n-heptane binary under a fixed pressure of 13.78 bars.


You are asked to determine the following quantities using the equation of state imposed at a pressure of 13.78 bar and at temperature K :

	Molar titers of the liquid x, and of the vapor y
	Molar enthalpy of liquid and vapor 
	Molar entropy of the liquid and the vapor 

Finally, plot your results in the attached tables and verify your results using the diagrams provided (T, x, y), (x, y) and (H, x, y).

We will take 
	h°(To)=0 with T_0=298.15 K
	s1°=310.1 J.mol.K-1 
	s2°=427.9 J.mol-1. K-1